Mods to give ConvertJob a dummy sample-sheet.
